Fix refetch before delete mutation
When the user deletes a problem, an alert pops up for them to confirm
the action. But when the user confirms, the window focus changes and
that triggers a refetch for the /api/problems query.

A mutation is triggered after the user confirmation, and refetches are
disabled if there is any mutation ocurruing, but the focus change
happens just before the trigger.

The current solution is to disable the focus before the alert pops up
and re-enable it after the mutate call.