Remove Next.js
Next was removed because no feature from it was being used and it just added an extra layer of complexity to configurations.

Config files are also deleted/changed to correspond as closely as possible as the one in the official react boilerplate code for typescript.

Pages are now broken because of next/router. This will be fixed later when it is replaced by react-router.